리로드 상태 추가 & 수정 중

+++ Underground_Gamers/Assets/Game Scene Assets/Scripts/AttackData/AttackDefinition.cs
@@ -16,7 +16,7 @@ public class AttackDefinition : ScriptableObject
     public float reloadCooldown;
 
     [Tooltip("장전 수")]
-    public float chargeCount;
+    public int chargeCount;
     [Tooltip("명중률")]
     public float accuracyRate;
     [Tooltip("집탄율")]
